Restructure project layout.
