+++ AdvancedDLSupport.Example/AdvancedDLSupport.Example.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AdvancedDLSupport\AdvancedDLSupport.csproj" />
+  </ItemGroup>
+  <Target Name="CompileCProject" AfterTargets="AfterBuild">
+    <exec Command="clang -std=c99 -shared -fPIC -oc\libDemo.so c\src\Example.c" />
+    <Copy SourceFiles="c\libDemo.so" DestinationFolder="$(OutDir)" />
+  </Target>
+</Project>
